+++ tests/HasAttributesTest.php
@@ -197,12 +197,10 @@ final class HasAttributesTest extends TestCase
             use HasAttributes;
         };
 
-        $closure = static fn(): string => 'resolved-value';
-
-        $instance = $instance->setAttribute('test', $closure);
+        $instance = $instance->setAttribute('test', static fn(): string => 'resolved-value');
 
         self::assertSame(
-            ['test' => $closure],
+            ['test' => 'resolved-value'],
             $instance->getAttributes(),
             "Should return the correct 'data-test' attribute after setting it.",
         );
@@ -222,7 +220,7 @@ final class HasAttributesTest extends TestCase
             "Should return 'null' after setting the attribute to 'null'.",
         );
         self::assertSame(
-            [],
+            ['id' => null],
             $instance->getAttributes(),
             'Should remove the attribute key when set to null.',
         );